Table backgrounds should use the visibility of the cell, not of the col/colgroup/row/rowgroup
Per spec at https://drafts.csswg.org/css-tables-3/#drawing-cell-backgrounds cell backgrounds (including ones arising from rows/cols) are supposed to only be drawn only on visible cells.  That means they're not drawn on cells with "visibility:hidden" or on empty cells when "empty-cells: hide" is in play.

part 1.  Add a method on nsTableCellFrame to determine whether it should paint borders and backgrounds.

part 2.  Table backgrounds should be drawn or not based on the visibility of the relevant cell, not of the column/row/colgroup/rowgroup they come from.